Add comments to tests
