+++ tests/non_interactive_mode_integration_tests/test_app_invalid_scenarious.py
@@ -1,4 +1,3 @@
-import subprocess
 import os
 from test_utils.test_utils import (
     run_package_to_image_placer,
@@ -11,7 +10,7 @@ from test_utils.test_utils import (
 
 
 def test_write_one_big_package(package_to_image_placer_binary):
-    """TODO"""
+    """Tests if the package_to_image_placer will fail when the package is bigger than the image"""
     config = "test_data/test_config.json"
     img_in = "test_data/test_img.img.in"
     img_out = "test_data/test_img_out.img"
@@ -32,6 +31,7 @@ def test_write_one_big_package(package_to_image_placer_binary):
 
 
 def test_write_two_big_packages(package_to_image_placer_binary):
+    """Tests if the package_to_image_placer will fail when the sum of packages is bigger than the image"""
     config = "test_data/test_config.json"
     img_in = "test_data/test_img.img.in"
     img_out = "test_data/test_img_out.img"
@@ -54,7 +54,7 @@ def test_write_two_big_packages(package_to_image_placer_binary):
 
 
 def test_write_one_package_to_small_partition(package_to_image_placer_binary):
-    """TODO"""
+    """Tests if the package_to_image_placer will fail when the package is bigger than the partition"""
     config = "test_data/test_config.json"
     img_in = "test_data/test_img.img.in"
     img_out = "test_data/test_img_out.img"
@@ -75,7 +75,7 @@ def test_write_one_package_to_small_partition(package_to_image_placer_binary):
 
 
 def test_write_multiple_packages_to_partition(package_to_image_placer_binary):
-    """TODO"""
+    """Tests if the package_to_image_placer will fail when the sum of packages is bigger than the partition"""
     config = "test_data/test_config.json"
     img_in = "test_data/test_img.img.in"
     img_out = "test_data/test_img_out.img"
@@ -101,41 +101,3 @@ def test_write_multiple_packages_to_partition(package_to_image_placer_binary):
 
     assert result.returncode == 1
     assert not os.path.exists(img_out)
-
-
-def test_double_write_without_override(package_to_image_placer_binary):
-    """TODO"""
-    config_1 = "test_data/test_config_1.json"
-    config_2 = "test_data/test_config_2.json"
-    img_in = "test_data/test_img.img.in"
-    img_out_1 = "test_data/test_img_out_1.img"
-    img_out_2 = "test_data/test_img_out_2.img"
-    package = "test_data/normal_package"
-    package_zip = package + ".zip"
-    partitions = [1]
-
-    create_test_package(package, "10KB")
-    create_image(img_in, "10MB", 1)
-    make_image_mountable(img_in)
-
-    create_config(config_1, img_in, img_out_1, [package_zip], partitions)
-    create_config(config_2, img_out_1, img_out_2, [package_zip], partitions)
-
-    result = run_package_to_image_placer(package_to_image_placer_binary, config=config_1)
-
-    assert result.returncode == 0
-    assert inspect_image(config_1)
-
-    # crete file that will be deleted because of the second run that will fail
-    with open(img_out_2, "wb") as f:
-        f.write(b"0")
-    assert os.path.exists(img_out_2)
-
-    result = run_package_to_image_placer(package_to_image_placer_binary, config=config_2)
-
-    assert result.returncode == 1
-
-    assert not os.path.exists(img_out_2)
-
-    # check if the source image was not damaged during second run
-    assert inspect_image(config_1)
